repository_manager/requirements.txt
argparse
PyYAML
GitPython
pandas
tabulate
datetime
pytz
tzlocal
termcolor

# requirements for mkdocs
mkdocs
mkdocs-material
pymdown-extensions
fontawesome_markdown
markdown>=3.2.1
python-markdown-math
plantuml-markdown